refactor: update import path for HelpAccordion to improve module structure
